fixed building issue
