+++ src/main/deploy/choreo/test.chor
@@ -1,6 +1,6 @@
 {
  "name":"test",
- "version":"v2025.0.0",
+ "version":1,
  "type":"Swerve",
  "variables":{
   "expressions":{},
@@ -32,8 +32,8 @@
    "val":68.0388555
   },
   "inertia":{
-   "exp":"6 kg m ^ 2",
-   "val":6.0
+   "exp":"1 kg m ^ 2",
+   "val":1.0
   },
   "gearing":{
    "exp":"6.5",
@@ -44,12 +44,16 @@
    "val":0.0508
   },
   "vmax":{
-   "exp":"502.65482457436696 rad / s",
-   "val":502.65482457436696
+   "exp":"6000 RPM",
+   "val":628.3185307179587
   },
   "tmax":{
-   "exp":"0.8717213114754098 N m",
-   "val":0.8717213114754098
+   "exp":"1.2 N * m",
+   "val":1.2
+  },
+  "cof":{
+   "exp":"1.5",
+   "val":1.5
   },
   "bumper":{
    "front":{
