+++ src/main/kotlin/frc/team449/Main.kt
@@ -5,5 +5,5 @@ package frc.team449
 import edu.wpi.first.wpilibj.RobotBase
 
 fun main() {
-  RobotBase.startRobot { RobotLoop() }
+  RobotBase.startRobot { Robot() }
 }
