+++ src/main/kotlin/frc/team449/Robot.kt
@@ -1,44 +1,125 @@
 package frc.team449
 
-import edu.wpi.first.wpilibj.PowerDistribution
-import edu.wpi.first.wpilibj2.command.button.CommandXboxController
-import frc.team449.subsystems.RobotConstants
-import frc.team449.subsystems.drive.swerve.SwerveDrive
-import frc.team449.subsystems.drive.swerve.SwerveOrthogonalCommand
-import frc.team449.subsystems.light.Light.Companion.createLight
-import frc.team449.subsystems.vision.PoseSubsystem.Companion.createPoseSubsystem
-import frc.team449.system.AHRS
+import com.ctre.phoenix6.SignalLogger
+import edu.wpi.first.hal.FRCNetComm
+import edu.wpi.first.hal.HAL
+import edu.wpi.first.wpilibj.*
+import edu.wpi.first.wpilibj.RobotBase
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard
+import edu.wpi.first.wpilibj2.command.CommandScheduler
+import edu.wpi.first.wpilibj2.command.InstantCommand
+import edu.wpi.first.wpilibj2.command.button.RobotModeTriggers
+import frc.team449.auto.Routines
+import frc.team449.commands.light.BlairChasing
+import frc.team449.commands.light.Rainbow
+import frc.team449.subsystems.drive.swerve.SwerveSim
+import frc.team449.subsystems.vision.VisionConstants
 import monologue.Annotations.Log
 import monologue.Logged
+import monologue.Monologue
+import org.littletonrobotics.urcl.URCL
 
-class Robot : RobotBase(), Logged {
-
-  val driveController = CommandXboxController(0)
-
-  val mechController = CommandXboxController(1)
-
-  val ahrs = AHRS()
-
-  // Instantiate/declare PDP and other stuff here
-  @Log.NT
-  override val powerDistribution: PowerDistribution = PowerDistribution(
-    RobotConstants.PDH_CAN,
-    PowerDistribution.ModuleType.kRev
-  )
+/** The main class of the robot, constructs all the subsystems
+ * and initializes default commands . */
+class Robot : TimedRobot(), Logged {
 
   @Log.NT
-  override val drive = SwerveDrive.createSwerveKraken(field)
+  private val robot = RobotContainer()
 
   @Log.NT
-  override val poseSubsystem = createPoseSubsystem(ahrs, drive, field)
+  private val field = robot.field
 
-  @Log.NT
-  override val driveCommand = SwerveOrthogonalCommand(drive, poseSubsystem, driveController.hid)
+  private val controllerBinder = ControllerBindings(robot.driveController, robot.mechController, robot)
 
-  val light = createLight()
+  override fun robotInit() {
+    // Yes this should be a print statement, it's useful to know that robotInit started.
+    println("Started robotInit.")
 
-  /** Example for using infrared sensors
-   * @Log.NT
-   * val infrared = DigitalInput(RobotConstants.IR_CHANNEL)
-   */
+    SignalLogger.setPath("/media/sda1/ctre-logs/")
+    SignalLogger.start()
+
+    HAL.report(FRCNetComm.tResourceType.kResourceType_Language, FRCNetComm.tInstances.kLanguage_Kotlin)
+
+    if (RobotBase.isSimulation()) {
+      // Don't complain about joysticks if there aren't going to be any
+      DriverStation.silenceJoystickConnectionWarning(true)
+//      val instance = NetworkTableInstance.getDefault()
+//      instance.stopServer()
+//      instance.startClient4("localhost")
+    }
+
+    /** Example Quad Calibration
+     QuadCalibration(robot.pivot).ignoringDisable(true).schedule()
+     */
+
+    println("Generating Auto Routines : ${Timer.getFPGATimestamp()}")
+    val routines = Routines(robot)
+    routines.addOptions(robot.autoChooser)
+    SmartDashboard.putData(robot.autoChooser)
+    RobotModeTriggers.autonomous().whileTrue(robot.autoChooser.selectedCommandScheduler())
+    println("DONE Generating Auto Routines : ${Timer.getFPGATimestamp()}")
+
+    SmartDashboard.putData("Command Scheduler", CommandScheduler.getInstance())
+
+    robot.light.defaultCommand = BlairChasing(robot.light)
+
+    controllerBinder.bindButtons()
+
+    DriverStation.startDataLog(DataLogManager.getLog())
+    Monologue.setupMonologue(this, "/Monologuing", false, false)
+
+    URCL.start()
+  }
+
+  override fun driverStationConnected() {
+    Monologue.setFileOnly(DriverStation.isFMSAttached())
+  }
+
+  override fun robotPeriodic() {
+    CommandScheduler.getInstance().run()
+
+    robot.field.robotPose = robot.poseSubsystem.pose
+
+    robot.field.getObject("bumpers").pose = robot.poseSubsystem.pose
+
+    Monologue.updateAll()
+  }
+
+  override fun autonomousInit() {}
+
+  override fun autonomousPeriodic() {}
+
+  override fun teleopInit() {
+    (robot.light.currentCommand ?: InstantCommand()).cancel()
+
+    robot.drive.defaultCommand = robot.driveCommand
+  }
+
+  override fun teleopPeriodic() {
+  }
+
+  override fun disabledInit() {
+    robot.drive.stop()
+
+    (robot.light.currentCommand ?: InstantCommand()).cancel()
+    Rainbow(robot.light).schedule()
+  }
+
+  override fun disabledPeriodic() {}
+
+  override fun testInit() {}
+
+  override fun testPeriodic() {}
+
+  override fun simulationInit() {}
+
+  override fun simulationPeriodic() {
+    robot.drive as SwerveSim
+
+    VisionConstants.ESTIMATORS.forEach {
+      it.simulationPeriodic(robot.drive.odometryPose)
+    }
+
+    VisionConstants.VISION_SIM.debugField.getObject("EstimatedRobot").pose = robot.poseSubsystem.pose
+  }
 }
