+++ src/main/kotlin/frc/team449/commands/PIDPoseAlign.kt
@@ -1,4 +1,4 @@
-package frc.team449.auto.choreo
+package frc.team449.commands
 
 import edu.wpi.first.math.controller.PIDController
 import edu.wpi.first.math.geometry.Pose2d
